refactor: consolidate common interfaces into shared types
- Move ProfileExportResult and AppSettings interfaces to shared/types.ts
- Remove duplicate StyleOptions interface from FineTuneControls.tsx
- Update all imports to use shared types consistently
- Improve type organization: services/types.ts for AI-specific, shared/types.ts for app-wide interfaces
